Move `cvd/unittests/parser` tests to `cvd/cli/parser`
Bug: b/416769281
