+++ base/cvd/cuttlefish/host/commands/cvd/cli/parser/flags_parser_test.cc
@@ -17,7 +17,7 @@
 #include <android-base/file.h>
 #include <gtest/gtest.h>
 
-#include "cuttlefish/host/commands/cvd/unittests/parser/test_common.h"
+#include "cuttlefish/host/commands/cvd/cli/parser/test_common.h"
 
 namespace cuttlefish {
