+++ base/cvd/cuttlefish/host/commands/cvd/cli/parser/instance/disk_configs_test.cc
@@ -16,7 +16,7 @@
 
 #include <gtest/gtest.h>
 
-#include "cuttlefish/host/commands/cvd/unittests/parser/test_common.h"
+#include "cuttlefish/host/commands/cvd/cli/parser/test_common.h"
 
 namespace cuttlefish {
