+++ base/cvd/cuttlefish/host/commands/cvd/cli/parser/instance/vm_configs_test.cc
@@ -21,7 +21,7 @@
 
 #include "cuttlefish/common/libs/utils/json.h"
 #include "cuttlefish/common/libs/utils/result_matchers.h"
-#include "cuttlefish/host/commands/cvd/unittests/parser/test_common.h"
+#include "cuttlefish/host/commands/cvd/cli/parser/test_common.h"
 
 namespace cuttlefish {
