--- base/cvd/cuttlefish/host/commands/cvd/unittests/parser/BUILD.bazel
@@ -1,44 +1,0 @@
-load("//:build_variables.bzl", "COPTS")
-load("//tools/lint:linters.bzl", "clang_tidy_test")
-
-package(
-    default_visibility = ["//:android_cuttlefish"],
-)
-
-cc_test(
-    name = "parser",
-    srcs = [
-        "configs_inheritance_test.cc",
-        "flags_parser_test.cc",
-        "instance/boot_configs_test.cc",
-        "instance/disk_configs_test.cc",
-        "instance/graphics_configs_test.cc",
-        "instance/vm_configs_test.cc",
-        "metrics_configs_test.cc",
-        "test_common.cc",
-        "test_common.h",
-    ],
-    copts = COPTS + [ "-Icuttlefish" ],
-    deps = [
-        "//cuttlefish/common/libs/utils",
-        "//cuttlefish/common/libs/utils:base64",
-        "//cuttlefish/common/libs/utils:result",
-        "//cuttlefish/common/libs/utils:result_matchers",
-        "//cuttlefish/host/commands/assemble_cvd:launch_cvd_cc_proto",
-        "//cuttlefish/host/commands/cvd/cli/parser",
-        "//cuttlefish/host/commands/cvd/cli/parser:configs_common",
-        "//cuttlefish/host/libs/config",
-        "//libbase",
-        "@googletest//:gtest",
-        "@googletest//:gtest_main",
-        "@jsoncpp",
-        "@protobuf",
-        "@protobuf//:differencer",
-    ],
-)
-
-clang_tidy_test(
-    name = "parser_clang_tidy",
-    srcs = [":parser"],
-    tags = ["clang_tidy", "clang-tidy"],
-)
